Add files via upload
A talking calculator has a built-in speech synthesizer that reads aloud each number, symbol, or operation key a user presses in Marathi.
Bharti Talking Calculator in Marathi, a talking calculator's main feature is the ability to talk in Marathi. The idea behind it is to make simple calculating tasks much more convenient and efficient to someone who cannot easily read a standard display. One does not need to give up any features and functions to get a calculator with Marathi talking capabilities